CvMDist(distrEx)	R Documentation

Generic function for the computation of the Cramer - von Mises distance of two distributions

Description

Generic function for the computation of the Cramer - von Mises distance d_{mu} of two distributions P and Q where the distributions are defined on a finite-dimensional Euclidean space (R^m, B^m) with B^m the Borel-sigma-algebra on R^m. The Cramer - von Mises distance is defined as

d_{mu}(P,Q)^2=int (P({y in R^m | y <= x})-Q({y in R^m | y <= x}))^2 mu(dx)

where <= is coordinatewise on R^m.

Usage


CvMDist(e1, e2, ...)
## S4 method for signature 'UnivariateDistribution,
##   UnivariateDistribution':
CvMDist(e1, e2, mu = e2, useApply = FALSE, ...)
## S4 method for signature 'numeric,
##   UnivariateDistribution':
CvMDist(e1, e2, mu = e2, ...)

Arguments

e1 	 object of class "Distribution" or class "numeric"

e2 	 object of class "Distribution"

... 	 further arguments to be used e.g. by E()

useApply 	 logical; to be passed to E()

mu 	 object of class "Distribution"; integration measure; defaulting to e2

Value

Cramer - von Mises distance of e1 and e2

Methods

e1 = "UnivariateDistribution", e2 = "UnivariateDistribution":

Cramer - von Mises distance of two univariate distributions.

e1 = "numeric", e2 = "UnivariateDistribution":

Cramer - von Mises distance between the empirical formed from a data set (e1) and a univariate distribution.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@itwm.fraunhofer.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

See Also

ContaminationSize, TotalVarDist, HellingerDist, KolmogorovDist, Distribution-class

Examples


CvMDist(Norm(), Gumbel())
CvMDist(Norm(), Gumbel(), mu = Norm())
CvMDist(Norm(), Td(10))
CvMDist(Norm(mean = 50, sd = sqrt(25)), Binom(size = 100))
CvMDist(Pois(10), Binom(size = 20)) 
CvMDist(rnorm(100),Norm())
CvMDist((rbinom(50, size = 20, prob = 0.5)-10)/sqrt(5), Norm())
CvMDist(rbinom(50, size = 20, prob = 0.5), Binom(size = 20, prob = 0.5))
CvMDist(rbinom(50, size = 20, prob = 0.5), Binom(size = 20, prob = 0.5), mu = Pois())

[Package distrEx version 2.0 Index]