vonMisesDist(distrEx)	R Documentation

Generic function for the computation of the vonMises distance of two distributions

Description

Generic function for the computation of the von Mises distance d_{mu} of two distributions P and Q where the distributions are defined on a finite-dimensional Euclidean space (R^m, B^m) with B^m the Borel-sigma-algebra on R^m. The von Mises distance is defined as

d_{mu}(P,Q)^2=int (P({y in R^m | y <= x})-Q({y in R^m | y <= x}))^2 mu(dx)

where <= is coordinatewise on R^m.

Usage


vonMisesDist(e1, e2, ...)
## S4 method for signature 'UnivariateDistribution,
##   UnivariateDistribution':
vonMisesDist(e1, e2, mu = e2)
## S4 method for signature 'numeric,
##   UnivariateDistribution':
vonMisesDist(e1, e2, mu = e2)

Arguments

e1 	 object of class "Distribution" or class "numeric"

e2 	 object of class "Distribution"

... 	 further arguments to be used in particular methods (not in package distrEx)

mu 	 object of class "Distribution"; integration measure; defaulting to e2

Value

von Mises distance of e1 and e2

Methods

e1 = "UnivariateDistribution", e2 = "UnivariateDistribution":

von Mises distance of two univariate distributions.

e1 = "numeric", e2 = "UnivariateDistribution":

von Mises distance between the empirical formed from a data set (e1) and a univariate distribution.

Author(s)

Matthias Kohl Matthias.Kohl@stamats.de,
Peter Ruckdeschel Peter.Ruckdeschel@uni-bayreuth.de

References

Rieder, H. (1994) Robust Asymptotic Statistics. New York: Springer.

See Also

ContaminationSize, TotalVarDist, HellingerDist, KolmogorovDist, Distribution-class

Examples


vonMisesDist(Norm(), Gumbel())
vonMisesDist(Norm(), Gumbel(), mu = Norm())
vonMisesDist(Norm(), Td(10))
vonMisesDist(Norm(mean = 50, sd = sqrt(25)), Binom(size = 100))
vonMisesDist(Pois(10), Binom(size = 20)) 
vonMisesDist(rnorm(100),Norm())
vonMisesDist((rbinom(50, size = 20, prob = 0.5)-10)/sqrt(5), Norm())
vonMisesDist(rbinom(50, size = 20, prob = 0.5), Binom(size = 20, prob = 0.5))
vonMisesDist(rbinom(50, size = 20, prob = 0.5), Binom(size = 20, prob = 0.5), mu = Pois())

[Package distrEx version 2.0 Index]